sigmavalue_backend/requirements.txt
Django
djangorestframework
pandas
openpyxl
python-dotenv
openai
gunicorn
